feat: color aggregate annotation tooltips
