+++ app/src/pages/project/metrics/AnnotationScoreTimeSeriesChart.tsx
@@ -12,6 +12,11 @@ import {
 
 import { Text } from "@phoenix/components";
 import {
+  AnnotationScoreText,
+  type AnnotationOptimizationConfig,
+  getPositiveOptimizationFromConfig,
+} from "@phoenix/components/annotation";
+import {
   ChartEmptyStateOverlay,
   ChartTooltip,
   ChartTooltipItem,
@@ -39,7 +44,14 @@ export type AnnotationScoreTimeSeriesDatum = {
   }>;
 };
 
-function TooltipContent({ active, payload, label }: TooltipContentProps) {
+function TooltipContent({
+  active,
+  payload,
+  label,
+  annotationConfigsByName,
+}: TooltipContentProps & {
+  annotationConfigsByName: ReadonlyMap<string, AnnotationOptimizationConfig>;
+}) {
   const { fullTimeFormatter } = useTimeFormatters();
   if (active && payload && payload.length) {
     return (
@@ -51,13 +63,24 @@ function TooltipContent({ active, payload, label }: TooltipContentProps) {
         )}
         {payload.map((entry, index) => {
           if (entry.value == null) return null;
+          const annotationName = String(entry.dataKey || "unknown");
+          const score = Number(entry.value);
           return (
             <ChartTooltipItem
               key={index}
               color={entry.color}
               shape="line"
-              name={String(entry.dataKey || "unknown")}
-              value={Number(entry.value).toFixed(2)}
+              name={annotationName}
+              value={
+                <AnnotationScoreText
+                  positiveOptimization={getPositiveOptimizationFromConfig({
+                    config: annotationConfigsByName.get(annotationName),
+                    score,
+                  })}
+                >
+                  {score.toFixed(2)}
+                </AnnotationScoreText>
+              }
             />
           );
         })}
@@ -96,12 +119,14 @@ export function AnnotationScoreTimeSeriesChart({
   scale,
   timeRange,
   onTimeRangeSelected,
+  annotationConfigsByName,
 }: {
   data: ReadonlyArray<AnnotationScoreTimeSeriesDatum>;
   names: ReadonlyArray<string>;
   scale: TimeBinScale;
   timeRange: TimeRange;
   onTimeRangeSelected?: (timeRange: TimeRange) => void;
+  annotationConfigsByName: ReadonlyMap<string, AnnotationOptimizationConfig>;
 }) {
   // Transform the data to have one property per annotation label
   const chartData = useMemo(
@@ -150,7 +175,15 @@ export function AnnotationScoreTimeSeriesChart({
                 tickFormatter={(x) => formatFloat(x)}
               />
               <CartesianGrid {...defaultCartesianGridProps} />
-              <Tooltip content={TooltipContent} {...defaultTooltipProps} />
+              <Tooltip
+                content={(props) => (
+                  <TooltipContent
+                    {...props}
+                    annotationConfigsByName={annotationConfigsByName}
+                  />
+                )}
+                {...defaultTooltipProps}
+              />
 
               {names.map((name) => {
                 return (
