+++ app/src/pages/project/metrics/ProjectAnnotationMetrics.tsx
@@ -1,7 +1,7 @@
 import { css } from "@emotion/react";
 import type { ReactNode } from "react";
 import { Suspense, useState } from "react";
-import { graphql, useFragment, useLazyLoadQuery } from "react-relay";
+import { graphql, useLazyLoadQuery } from "react-relay";
 
 import { Loading, Text } from "@phoenix/components";
 import {
@@ -32,7 +32,6 @@ import { getNonNoteAnnotationNames } from "../spanAnnotationUtils";
 import type { ProjectAnnotationMetricNamesSessionQuery } from "./__generated__/ProjectAnnotationMetricNamesSessionQuery.graphql";
 import type { ProjectAnnotationMetricNamesSpanQuery } from "./__generated__/ProjectAnnotationMetricNamesSpanQuery.graphql";
 import type { ProjectAnnotationMetricNamesTraceQuery } from "./__generated__/ProjectAnnotationMetricNamesTraceQuery.graphql";
-import type { ProjectAnnotationMetricsConfigFragment$key } from "./__generated__/ProjectAnnotationMetricsConfigFragment.graphql";
 import type { ProjectAnnotationMetricsSessionQuery } from "./__generated__/ProjectAnnotationMetricsSessionQuery.graphql";
 import type { ProjectAnnotationMetricsSpanQuery } from "./__generated__/ProjectAnnotationMetricsSpanQuery.graphql";
 import type { ProjectAnnotationMetricsTraceQuery } from "./__generated__/ProjectAnnotationMetricsTraceQuery.graphql";
@@ -41,6 +40,7 @@ import {
   PROJECT_METRICS_CHART_SYNC_ID,
   useMetricQueryFetchOptions,
 } from "./types";
+import { useProjectAnnotationConfigsByName } from "./useProjectAnnotationConfigsByName";
 
 type AnnotationMetricsData = ReadonlyArray<{
   readonly timestamp: string;
@@ -58,61 +58,6 @@ type ProjectAnnotationMetricsResult = {
   annotationSeries: AnnotationMetricsSeries[];
   annotationConfigsByName: ReadonlyMap<string, AnnotationOptimizationConfig>;
 };
-
-function useProjectAnnotationConfigsByName(
-  project: ProjectAnnotationMetricsConfigFragment$key
-): ReadonlyMap<string, AnnotationOptimizationConfig> {
-  const data = useFragment(
-    graphql`
-      fragment ProjectAnnotationMetricsConfigFragment on Project {
-        annotationConfigs(first: 100) {
-          edges {
-            config: node {
-              ... on AnnotationConfigBase {
-                name
-                annotationType
-              }
-              ... on CategoricalAnnotationConfig {
-                optimizationDirection
-                values {
-                  label
-                  score
-                }
-              }
-              ... on ContinuousAnnotationConfig {
-                optimizationDirection
-                lowerBound
-                upperBound
-              }
-              ... on FreeformAnnotationConfig {
-                optimizationDirection
-                threshold
-                lowerBound
-                upperBound
-              }
-            }
-          }
-        }
-      }
-    `,
-    project
-  );
-  const configsByName = new Map<string, AnnotationOptimizationConfig>();
-  data.annotationConfigs.edges.forEach(({ config }) => {
-    if (config.name == null || config.annotationType == null) {
-      return;
-    }
-    configsByName.set(config.name, {
-      annotationType: config.annotationType,
-      optimizationDirection: config.optimizationDirection,
-      lowerBound: config.lowerBound,
-      upperBound: config.upperBound,
-      threshold: config.threshold,
-      values: config.values,
-    });
-  });
-  return configsByName;
-}
 
 function getProjectAnnotationMetricsSeries({
   data,
